Create generic file validator
